+++ gazebo/rendering/ScenePrivate.hh
@@ -14,8 +14,8 @@
  * limitations under the License.
  *
 */
-#ifndef _SCENE_PRIVATE_HH_
-#define _SCENE_PRIVATE_HH_
+#ifndef _GAZEBO_RENDERING_SCENE_PRIVATE_HH_
+#define _GAZEBO_RENDERING_SCENE_PRIVATE_HH_
 
 #include <list>
 #include <map>
@@ -23,7 +23,6 @@
 #include <vector>
 
 #include <boost/unordered/unordered_map.hpp>
-#include <boost/thread/recursive_mutex.hpp>
 
 #include <sdf/sdf.hh>
 
@@ -207,10 +206,10 @@ namespace gazebo
       public: SkeletonPoseMsgs_L skeletonPoseMsgs;
 
       /// \brief Mutex to lock the various message buffers.
-      public: boost::mutex *receiveMutex;
+      public: std::mutex *receiveMutex;
 
       /// \brief Mutex to lock the pose message buffers.
-      public: boost::recursive_mutex poseMsgMutex;
+      public: std::recursive_mutex poseMsgMutex;
 
       /// \brief Communication Node
       public: transport::NodePtr node;